src/requirements.txt
azure-common
azure-core
azure-identity
azure-keyvault-secrets
certifi
cffi
chardet
cryptography
idna
msal
msrest
oauthlib
pycparser
PyJWT
requests
requests-oauthlib
schedule
selenium
six
webdriver-manager